The new VoteCounter has finally arrived!
This commit separates VoteCounter into three parts -- VoteClient, VoteServer,
and VoteCommon (library for both server and client), while still keeping the
old standalone VoteCounter as VoteCounterOld.
It has not been rigorously tested, but it will probably work, and there is a
lot of documentation to be added soon. Downloads will be available too.